Moved eslint to src and replaced react file extensions
