+++ app/src/components/Drawer/index.js
@@ -1,3 +1,3 @@
-import Drawer from "./Drawer";
+import Drawer from "./Drawer.jsx";
 
 export default Drawer;
