+++ app/src/components/Header/index.js
@@ -1,3 +1,3 @@
-import Header from "./Header";
+import Header from "./Header.jsx";
 
 export default Header;
