+++ app/src/routes/Home/index.js
@@ -1,4 +1,4 @@
-import HomeView from "./components/HomeView";
+import HomeView from "./components/HomeView.jsx";
 
 // Sync route definition
 export default {
